chore: move code around for exchange rates (first commit)
